Move CommandLineArguments type into lib
